+++ Windows/TreeView.d
@@ -142,9 +142,9 @@ imeth	int	gCheckValue()
 			char	cmd[100], ns[80];
 			object	ret;
 			int	res;
-			sprintf(cmd, "(%s (int->object %ld) (int->object %ld))",
+			sprintf(cmd, "(%s (int->object %lld) (int->object %lld))",
 				gFunctionName(SchemeClassSurrogate, (object)iAcf),
-				(long) self, (long) 0);
+				(long long) self, (long long) 0);
 			ret = gExecuteInNamespace(SchemeClassSurrogate,
 						  gNamespaceName(SchemeClassSurrogate, (object)iAcf, ns), 
 						  cmd);
@@ -156,7 +156,7 @@ imeth	int	gCheckValue()
 		} else if (JavaScriptClassSurrogate  &&  IsObj((object)iAcf)  &&  ClassOf(iAcf) == JavaScriptString) {
 			object	ret;
 			char	cmd[128];
-			sprintf(cmd, "%s(StringToObject(\"%ld\"), %ld)", gStringValue((object)iAcf), (long) self, (long) 0);
+			sprintf(cmd, "%s(StringToObject(\"%lld\"), %lld)", gStringValue((object)iAcf), (long long) self, (long long) 0);
 			ret = gExecuteString(JavaScriptClassSurrogate, cmd);
 			if (IsObj(ret)) {
 				if (r = ClassOf(ret) == String)
